+++ relevanceai/workflows/base.py
@@ -3,7 +3,6 @@
 from re import I
 import traceback
 from typing import Callable, Optional
-from relevanceai.interfaces import Dataset
 from doc_utils import DocUtils
 from uuid import uuid4
 
@@ -24,7 +23,7 @@ class Workflow(DocUtils):
 
     def fit_dataset(
         self,
-        dataset: Dataset,
+        dataset,
         input_field: str,
         output_field: str,
         filters: Optional[list] = None,
@@ -107,4 +106,6 @@ class Workflow(DocUtils):
         metadata = self.dataset.metadata
         if "workflows" not in metadata:
             metadata["workflows"] = []
-        self.dataset.metadata["workflows"].append(workflow_metadata)
+        workflows = metadata["workflows"]
+        workflows.append(workflow_metadata)
+        self.dataset.metadata["workflows"] = workflows