+++ relevanceai/workflows/cluster_ops/clusterops.py
@@ -26,7 +26,7 @@ import warnings
 import numpy as np
 
 from relevanceai.api.client import BatchAPIClient
-from typing import Union, List, Dict, Optional, Callable
+from typing import Union, List, Dict, Optional, Callable, Set
 from relevanceai.workflows.cluster_ops.cluster_base import (
     ClusterBase,
     CentroidClusterBase,
@@ -49,7 +49,8 @@ from relevanceai.dataset_interface import Dataset
 
 from relevanceai.package_utils.errors import NoDocumentsError
 from relevanceai.package_utils.version_decorators import beta
-
+from relevanceai.package_utils.concurrency import multiprocess
+from relevanceai.workflows.cluster_ops.cluster_evaluate import ClusterEvaluate
 from doc_utils import DocUtils
 
 
@@ -80,7 +81,7 @@ METRIC_DESCRIPTION = {
 }
 
 
-class ClusterOps(BatchAPIClient):
+class ClusterOps(ClusterEvaluate):
 
     _cred_fn = ".creds.json"
 
@@ -451,7 +452,7 @@ class ClusterOps(BatchAPIClient):
         filters = [] if filters is None else filters
 
         return self.datasets.cluster.centroids.list_closest_to_center(
-            dataset_id=self._retrieve_dataset_id(dataset),
+            dataset_id=self._check_dataset_id(dataset),
             vector_fields=self.vector_fields
             if vector_fields is None
             else vector_fields,
@@ -620,7 +621,7 @@ class ClusterOps(BatchAPIClient):
         filters = [] if filters is None else filters
 
         return self.services.cluster.aggregate(
-            dataset_id=self._retrieve_dataset_id(dataset),
+            dataset_id=self._check_dataset_id(dataset),
             vector_fields=self.vector_fields if not vector_fields else vector_fields,
             groupby=groupby,
             metrics=metrics,
@@ -685,7 +686,7 @@ class ClusterOps(BatchAPIClient):
 
         """
         return self.datasets.cluster.centroids.list_furthest_from_center(
-            dataset_id=self._retrieve_dataset_id(dataset),
+            dataset_id=self._check_dataset_id(dataset),
             vector_fields=self.vector_fields
             if vector_fields is None
             else vector_fields,
@@ -708,7 +709,7 @@ class ClusterOps(BatchAPIClient):
 
         return
 
-    def _retrieve_dataset_id(self, dataset: Optional[Union[str, Dataset]]) -> str:
+    def _check_dataset_id(self, dataset: Optional[Union[str, Dataset]] = None) -> str:
         """Helper method to get multiple dataset values"""
 
         if isinstance(dataset, Dataset):
@@ -761,7 +762,7 @@ class ClusterOps(BatchAPIClient):
         """
 
         results = self.services.cluster.centroids.insert(
-            dataset_id=self._retrieve_dataset_id(dataset),
+            dataset_id=self._check_dataset_id(dataset),
             cluster_centers=centroid_documents,
             vector_fields=self.vector_fields,
             alias=self.alias,
@@ -844,7 +845,7 @@ class ClusterOps(BatchAPIClient):
             base_url + "/services/cluster/centroids/delete",
             headers={"Authorization": self.project + ":" + self.api_key},
             params={
-                "dataset_id": self._retrieve_dataset_id(dataset),
+                "dataset_id": self._check_dataset_id(dataset),
                 "vector_field": vector_fields,
                 "alias": self.alias,
             },
@@ -1366,7 +1367,19 @@ class ClusterOps(BatchAPIClient):
                 + f"https://cloud.relevance.ai/dataset/{self.dataset_id}/deploy/recent/cluster"
             )
 
-    def unique_cluster_ids(self, alias: str = None, minimum_cluster_size: int = 10):
+    def _check_for_dataset_id(self):
+        if not hasattr(self, "dataset_id"):
+            raise ValueError(
+                "You are missing a dataset ID. Please set using the argument dataset_id='...'."
+            )
+
+    def unique_cluster_ids(
+        self,
+        alias: str = None,
+        minimum_cluster_size: int = 3,
+        dataset_id: str = None,
+        num_clusters: int = 1000,
+    ):
         """
         We call facets on our data, which looks a little like this:
 
@@ -1382,25 +1395,39 @@ class ClusterOps(BatchAPIClient):
         """
         # Mainly to be used for subclustering
         # Get the cluster alias
+        if dataset_id is None:
+            self._check_for_dataset_id()
+            dataset_id = self.dataset_id
+
         cluster_field = self._get_cluster_field_name(alias=alias)
 
-        facet_results = self.dataset.facets(
-            fields=[cluster_field],
-            page_size=int(self.config["data.max_clusters"]),
-            page=1,
-            asc=True,
-        )
-        all_cluster_ids = []
-        if "results" in facet_results:
-            facet_results = facet_results["results"]
-        if cluster_field not in facet_results:
-            raise ValueError(
-                f"No clusters with alias `{alias}`. Please check the schema."
+        # currently the logic for facets is that when it runs out of pages
+        # it just loops - therefore we need to store it in a simple hash
+        # and then add them to a list
+        all_cluster_ids: Set = set()
+
+        while len(all_cluster_ids) < num_clusters:
+            facet_results = self.datasets.facets(
+                dataset_id=dataset_id,
+                fields=[cluster_field],
+                page_size=int(self.config["data.max_clusters"]),
+                page=1,
+                asc=True,
             )
-        for facet in facet_results[cluster_field]:
-            if facet["frequency"] > minimum_cluster_size:
-                all_cluster_ids.append(facet[cluster_field])
-        return all_cluster_ids
+            if "results" in facet_results:
+                facet_results = facet_results["results"]
+            if cluster_field not in facet_results:
+                raise ValueError(
+                    f"No clusters with alias `{alias}`. Please check the schema."
+                )
+            for facet in facet_results[cluster_field]:
+                if facet["frequency"] > minimum_cluster_size:
+                    curr_len = len(all_cluster_ids)
+                    all_cluster_ids.add(facet[cluster_field])
+                    new_len = len(all_cluster_ids)
+                    if new_len == curr_len:
+                        return list(all_cluster_ids)
+        return list(all_cluster_ids)
 
     @track
     def fit_dataset(
@@ -2259,3 +2286,165 @@ class ClusterOps(BatchAPIClient):
         return self._report.internal_report
 
     report = internal_report
+
+    def operate(self, field: str, func: Callable, output_field: Optional[str] = None):
+        """
+        Run an function per cluster.
+
+        Parameters
+        ------------
+
+        field: str
+            The field to operate on
+        func: Callable
+            The function to run on all the values once they are received.
+            It should take in a list of values
+        output_field: Optional[str]
+            Outputs for every document in the cluster
+
+        Example
+        ---------
+
+        .. code-block::
+
+            import numpy as np
+            def vector_mean(vectors):
+                return np.mean(vectors, axis=0)
+            cluster_centroids = cluster_ops.operate(
+                field="review_sentence_answer_use_vector_",
+                func=vector_mean
+            )
+
+        """
+        # Run a function on each cluster
+        output: Dict = {}
+        cluster_ids = self.unique_cluster_ids()
+        for cluster_id in tqdm(cluster_ids):
+            self._operate(cluster_id, field, output, func)
+        if output_field is not None:
+            self.update_documents_within_clusters(output, output_field)
+        return output
+
+    def _operate(self, cluster_id: str, field: str, output: dict, func: Callable):
+        """
+        Internal function for operations
+        """
+        cluster_field = self._get_cluster_field_name()
+        # TODO; change this to fetch all documents
+        documents = self.datasets.documents.get_where(
+            self.dataset_id,
+            filters=[
+                {
+                    "field": cluster_field,
+                    "filter_type": "exact_match",
+                    "condition": "==",
+                    "condition_value": cluster_id,
+                },
+                {
+                    "field": field,
+                    "filter_type": "exists",
+                    "condition": ">=",
+                    "condition_value": " ",
+                },
+            ],
+            select_fields=[field, cluster_field],
+            page_size=9999,
+        )
+        # get the field across each
+        arr = self.get_field_across_documents(field, documents["documents"])
+        output[cluster_id] = func(arr)
+
+    def create_centroids(
+        self, vector_fields: List[str], operation: Optional[Callable] = None
+    ):
+        """
+        Create centroids if there are none. The default operation is to take the centroid
+        of each vector.
+        An alternative function can be provided provided.
+
+        Example of the operation in question for mean:
+
+        .. code-block::
+
+            def vector_mean(vectors):
+                return np.mean(vectors, axis=0)
+
+            clusterops.create_centroids(["sample_vector_"], operation=vector_mean)
+
+        """
+        if len(vector_fields) > 1:
+            raise ValueError("currently do not support more than 1 vector field.")
+        vector_field: str = vector_fields[0]
+
+        def vector_mean(vectors):
+            return np.mean(vectors, axis=0)
+
+        if operation == "mean" or operation is None:
+            operation = vector_mean
+
+        cluster_centroids = self.operate(field=vector_field, func=operation)  # type: ignore
+        centroid_docs = []
+        for k, v in cluster_centroids.items():
+            centroid_docs.append({"_id": str(k), vector_field: v.tolist()})
+        return self.insert_centroid_documents(centroid_docs)
+
+    def _get_filter_for_cluster(self, cluster_id):
+        cluster_field = self._get_cluster_field_name()
+        filters = [
+            {
+                "field": cluster_field,
+                "filter_type": "exact_match",
+                "condition": "==",
+                "condition_value": cluster_id,
+            }
+        ]
+        return filters
+
+    def update_documents_within_cluster(self, cluster_id: str, update: dict):
+        """
+        Update all the documents within a cluster
+        """
+        cluster_filter = self._get_filter_for_cluster(cluster_id)
+        result = self.datasets.documents.update_where(
+            dataset_id=self.dataset_id, update=update, filters=cluster_filter
+        )
+        return result
+
+    def update_documents_within_clusters(
+        self, cluster_id_and_updates: dict, output_field: str = None
+    ):
+        """
+        Takes the cluster ids and updates and updates them accordingly
+
+        Example
+        ---------
+
+        .. code-block::
+
+            # Let us take the operator results
+            operator_results = clusterops.operate(...)
+            clusterops.update_documents_within_clusters(
+                operator_results, output_field="centroid_vector_"
+            )
+        """
+        results = []
+        for cluster_id, update_value in cluster_id_and_updates.items():
+            if type(update_value) != dict:
+                if output_field is None:
+                    raise ValueError(
+                        """
+                        As update value is not a dictionary, you will need to
+                        specify `output_field=`.
+                    """
+                    )
+                update: dict = {}
+                self.set_field(output_field, update, update_value)
+            else:
+                update = update_value
+
+            cluster_filter = self._get_filter_for_cluster(cluster_id)
+            result = self.datasets.documents.update_where(
+                dataset_id=self.dataset_id, update=update, filters=cluster_filter
+            )
+            results.append(result)
+        return results